Add HttpBasic auth to WebEndpoints

+++ src/main/java/cz/cvut/fit/nebesluk/tjv_semestral_client/apiClient/ClientClient.java
@@ -2,7 +2,10 @@ package cz.cvut.fit.nebesluk.tjv_semestral_client.apiClient;
 
 import cz.cvut.fit.nebesluk.tjv_semestral_client.dto.client.ClientDto;
 import cz.cvut.fit.nebesluk.tjv_semestral_client.dto.client.NewClientDto;
+import org.glassfish.jersey.client.authentication.HttpAuthenticationFeature;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 import javax.ws.rs.client.ClientBuilder;
@@ -50,6 +53,14 @@ public class ClientClient {
     }
 
     public Optional<ClientDto> putClient(Long id, NewClientDto dto){
+
+        var auth = SecurityContextHolder.getContext().getAuthentication();
+        if(auth instanceof AnonymousAuthenticationToken){
+            return Optional.empty();
+        }
+        HttpAuthenticationFeature basic = HttpAuthenticationFeature.basic(auth.getName(),auth.getCredentials().toString());
+
+
         var res =  specificClientEndpointTemplate.resolveTemplate("id",id).request(MediaType.APPLICATION_JSON_TYPE)
                 .put(Entity.entity(dto,MediaType.APPLICATION_JSON_TYPE));
 
@@ -61,6 +72,13 @@ public class ClientClient {
     }
 
     public void deleteClient(Long id){
+
+        var auth = SecurityContextHolder.getContext().getAuthentication();
+        if(auth instanceof AnonymousAuthenticationToken){
+            throw new RuntimeException("Login is required");
+        }
+        HttpAuthenticationFeature basic = HttpAuthenticationFeature.basic(auth.getName(),auth.getCredentials().toString());
+
         var res = specificClientEndpointTemplate.resolveTemplate("id",id)
                 .request().delete();
 
@@ -69,8 +87,11 @@ public class ClientClient {
         }
     }
 
-    public boolean CheckLogin(String baseAuthorization){
-        return clientAuthEndpoint.request().header("Authorization",baseAuthorization)
+    public boolean CheckLogin(String username,String password){
+
+        HttpAuthenticationFeature basic = HttpAuthenticationFeature.basic(username,password);
+
+        return clientAuthEndpoint.register(basic).request()
                 .get(boolean.class);
     }
